+++ CHANGELOG.md
@@ -3,12 +3,12 @@
 - MAJOR: remove the deprecated static method getCurrentTimeOfDayDimValue
 - MAJOR: provide the ES6 package as the default, and the ES5 one as a legacy package at appgrid/legacy
 - MINOR: basic robot detection with a default deviceId for them all ?
-- MINOR: have extra configuration params given go through the middleware so they are used in the factory (that will allow setting the gid when using the middleware, for instance)
 
 # Unreleased
 
 - MINOR: getEntries can now take a typeAlias parameter
 - MINOR: Added the new method to add Application Logs in batch (sendLogs)
+- MINOR: have extra configuration params given go through the middleware so they are used in the factory (log, gid, etc)
 - PATCH: The sendLog method was not sending log dimensions properly.
 - PATCH: The sendLog method's metadata param can be anything (not only an array as documented).
 - PATCH: Changed the AppGrid REST API link to its new location